Consolidate Rose offerings and rename Programs to Offerings
- Add aura reading levels (1-3) and journey overview to The Rose page,
  so it serves as the single source for all rose and aura descriptions
- Remove the old Offerings page (was duplicate of Rose page content)
- Rename Programs page to Offerings, now at /offerings route
- Delete /programs route
- Update navigation, footer, and all cross-references site-wide
